+++ examples/intsum/SingleStrict.hs
@@ -7,5 +7,5 @@ module SingleStrict (
 import Data.Int
 import qualified Data.Vector.Unboxed as U
 
-sumFrom1To :: U.Vector Int32 -> Int32
+sumFrom1To :: U.Vector Int64 -> Int64
 sumFrom1To v = U.foldl' (+) 0 v
